pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leo-physick-tgn"
version = "0.1.0"
description = "Physics-Guided Temporal Graph Learning Framework for Handover Management in Ultra-Dense LEO Networks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1",
  "pyyaml>=6.0",
  "numpy>=1.24",
  "skyfield>=1.45",
  "sgp4>=2.23",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
